The app has been made pretty. Global styles were moved to styles.css. Featured recipes are no longer hard coded.
