+++ src/app/header/header.component.html
@@ -1,4 +1,4 @@
-<nav class="navbar navbar-default">
+<nav class="navbar navbar-default metallic">
   <div class="container-fluid">
     <div class="navbar-header">
       <a routerLink="/" class="navbar-brand">Recipe Book</a>
@@ -8,6 +8,7 @@
         <li routerLinkActive="app-active"><a routerLink="/recipes">Recipes</a></li>
         <li routerLinkActive="app-active"><a routerLink="/featured-recipes">Featured Recipes</a></li>
         <li routerLinkActive="app-active"><a routerLink="/shopping-list">Shopping List</a></li>
+        <li routerLinkActive="app-active"><a routerLink="/hanoi">Train Your Brain!</a></li>
       </ul>
       <ul class="nav navbar-nav navbar-right">
         <li *ngIf="recipesUnsaved || shoppingListUnsaved"><em>Changes Unsaved!</em></li>
