+++ src/app/recipes/featured-recipes/f-recipe-detail/f-recipe-detail.component.css
@@ -3,13 +3,9 @@ img {
 }
 
 .root {
-  overflow-y: scroll;
-  overflow-x: hidden;
   width: 100%;
   max-height: 85vh;
   color: black;
-  background-color: #fbb034;
-  background-image: linear-gradient(315deg, #fbb034 0%, #ffdd00 74%);
   border-radius: 5px;
   padding: 10px;
 }
@@ -23,9 +19,13 @@ img {
 }
 
 .ing-list>li {
-  background: transparent !important;
+  box-shadow: 3px 3px 5px;
   text-align: center;
   border: none;
+}
+
+.ing-list>li:hover {
+  margin-left: 10px;
 }
 
 hr {
