+++ src/app/recipes/featured-recipes/f-recipe-detail/f-recipe-detail.component.html
@@ -1,4 +1,4 @@
-<div class='root'>
+<div class='root metallic scrollable'>
 <div class="row app-float-top">
     <div class="col-xs-12 centered">
       <img [src]="featuredRecipe.imagePath" [alt]="featuredRecipe.name" class="image-responsive strong-border">
@@ -12,7 +12,7 @@
   <div class="row">
     <div class="col-xs-12 centered">
       <div class="btn-group" appDropdown>
-        <button type="button" class="btn btn-primary dropdown-toggle">Manage Recipe <span class="caret"></span></button>
+        <button type="button" class="btn dropdown-toggle functional">Manage Recipe <span class="caret"></span></button>
         <ul class="dropdown-menu">
           <li><a style="cursor: pointer;" (click)="addToShoppingList()">To Shopping List</a></li>
           <li><a style="cursor: pointer;" (click)="addToRecipes()">Add to Recipes</a></li>
@@ -31,8 +31,8 @@
   <div class="row">
     <div class="col-xs-12">
       <strong>Ingredients:</strong>
-      <ul class="list-group ing-list">
-        <li class="list-group-item"
+      <ul style="margin-top: 10px;" class="list-group ing-list">
+        <li class="list-group-item metallic-reverse"
             *ngFor="let ingredient of featuredRecipe.ingredients">-&nbsp;&nbsp;&nbsp;{{ ingredient.name }} - {{ ingredient.amount }}</li>
       </ul>
       <p *ngIf="timesAddedToShoppingList > 0"><em>Added to Shopping List! (x{{timesAddedToShoppingList}})</em><span style="cursor: pointer; margin-left: 5px; font-size: xx-large; color: green; text-anchor:end;" (click)="addToShoppingList()">+</span></p>
