+++ src/app/recipes/featured-recipes/f-recipe-list/f-recipe-list.component.css
@@ -1,0 +1,3 @@
+.root {
+  max-height: 85vh;
+}
