+++ src/app/recipes/featured-recipes/f-recipe-list/f-recipe-list.component.html
@@ -1,4 +1,4 @@
-<div class="row">
+<div class="row root scrollable">
   <div class="col-xs-12" *ngFor="let recipe of featuredRecipes; let id = index">
     <app-recipe-item [recipe]="recipe" [index]="id"></app-recipe-item>
   </div>
