+++ src/app/recipes/recipe-detail/recipe-detail.component.html
@@ -1,4 +1,4 @@
-<div class='root'>
+<div class='root metallic scrollable'>
   <div class="row app-float-top">
       <div class="col-xs-12 centered">
         <img [src]="recipe.imagePath" [alt]="recipe.name" class="image-responsive strong-border">
@@ -32,11 +32,10 @@
       <div class="col-xs-12">
         <strong>Ingredients:</strong>
         <ul class="list-group ing-list">
-          <li class="list-group-item"
+          <li class="list-group-item metallic-reverse"
               *ngFor="let ingredient of recipe.ingredients">-&nbsp;&nbsp;&nbsp;{{ ingredient.name }} - {{ ingredient.amount }}</li>
         </ul>
         <p *ngIf="timesAddedToShoppingList > 0"><em>Added to Shopping List! (x{{timesAddedToShoppingList}})</em><span style="cursor: pointer; margin-left: 5px; font-size: xx-large; color: green; text-anchor:end;" (click)="addToShoppingList()">+</span></p>
       </div>
     </div>
   </div>
-
