+++ src/app/recipes/recipe-edit/recipe-edit.component.html
@@ -1,96 +1,99 @@
-<div class="row">
+<div class="row root">
   <div class="col-xs-12">
     <form [formGroup]="recipeForm" (ngSubmit)="onSubmit()">
       <div class="row">
-        <div class="col-xs-12">
+        <div class="col-xs-12 centered">
           <button [disabled]="recipeForm.invalid" type="submit" class="btn btn-success">Save</button>
           <button routerLink="../" type="button" class="btn btn-danger">Cancel</button>
         </div>
       </div>
-      <div class="row">
-        <div class="col-xs-12">
-          <div class="form-group">
-            <label for="name">Name</label>
-            <input
-              type="text"
-              id="name"
-              class="form-control"
-              formControlName="name"
-            >
-          </div>
-        </div>
-      </div>
-      <div class="row">
-        <div class="col-xs-12">
-          <div class="form-group">
-            <label for="imagePath">Image URL</label>
-            <input
-              type="text"
-              id="imagePath"
-              class="form-control"
-              formControlName="imagePath"
-              #imageInput
-            >
-          </div>
-        </div>
-      </div>
-      <div class="row">
-        <div class="col-xs-12">
-          <img [src]="imageInput.value" class="image-responsive">
-        </div>
-      </div>
-      <div class="row">
-        <div class="col-xs-12">
-          <label for="description"></label>
-          <textarea
-            type="text"
-            id="description"
-            class="form-control"
-            rows="6"
-            formControlName="description"
-          ></textarea>
-        </div>
-      </div>
-      <div class="row">
-        <div class="col-xs-12" formArrayName="ingredients">
-          <div
-            class="row"
-            *ngFor="let control of getFormControls(); let i = index"
-            [formGroupName]="i"
-            style="margin-top: 10px;"
-          >
-            <div class="col-xs-8">
+      <div class="content scrollable">
+        <div class="row">
+          <div class="col-xs-12">
+            <div class="form-group">
+              <label for="name">Name</label>
               <input
                 type="text"
+                id="name"
                 class="form-control"
                 formControlName="name"
               >
             </div>
-            <div class="col-xs-2">
+          </div>
+        </div>
+        <div class="row">
+          <div class="col-xs-12">
+            <div class="form-group">
+              <label for="imagePath">Image URL</label>
               <input
-                type="number"
+                type="text"
+                id="imagePath"
                 class="form-control"
-                formControlName="amount"
+                formControlName="imagePath"
+                #imageInput
               >
             </div>
-            <div class="col-xs-2">
-              <button
-                (click)="onDeleteIngredientFromRecipe(i)"
-                type="button"
-                class="btn btn-danger"
-              >X</button>
-            </div>
           </div>
-          <hr>
-          <div class="row">
-            <div class="col-xs-12">
-              <button
-                class="btn btn-success"
-                (click)="onAddIngredient()"
-                type="button"
-              >Add Ingredient</button>
-            </div>
+        </div>
+        <div class="row">
+          <div class="col-xs-12 centered">
+            <img [src]="imageInput.value" class="image-responsive">
           </div>
+        </div>
+        <div class="row">
+          <div class="col-xs-12">
+            <label for="description"></label>
+            <textarea
+              type="text"
+              id="description"
+              class="form-control"
+              rows="6"
+              formControlName="description"
+            ></textarea>
+          </div>
+        </div>
+        <div class="row">
+          <div class="col-xs-12" formArrayName="ingredients">
+            <div class="ing-list scrollable">
+              <div
+                class="row"
+                *ngFor="let control of getFormControls(); let i = index"
+                [formGroupName]="i"
+                style="margin-top: 10px;"
+              >
+                <div class="col-xs-8">
+                  <input
+                    type="text"
+                    class="form-control"
+                    formControlName="name"
+                  >
+                </div>
+                <div class="col-xs-2">
+                  <input
+                    type="number"
+                    class="form-control"
+                    formControlName="amount"
+                  >
+                </div>
+                <div class="col-xs-2">
+            </div>
+                <button
+                  (click)="onDeleteIngredientFromRecipe(i)"
+                  type="button"
+                  class="btn btn-danger"
+                >X</button>
+              </div>
+            </div>
+      </div>
+    </div>
+    <div class="row">
+      <div class="col-xs-12">
+        <button
+          class="btn btn-success"
+          (click)="onAddIngredient()"
+          type="button"
+        >Add Ingredient</button>
+      </div>
         </div>
       </div>
     </form>
