+++ src/app/recipes/recipe-item/recipe-item.component.css
@@ -3,8 +3,6 @@
 }
 
 a {
-  background-color: #fbb034;
-  background-image: linear-gradient(315deg, #fbb034 0%, #ffdd00 74%);
   border: none;
   opacity: 50%;
 }
