+++ src/app/recipes/recipe-item/recipe-item.component.html
@@ -1,4 +1,4 @@
-<a [routerLink]="[index]" routerLinkActive="app-active" class="list-group-item clearfix" style="cursor: pointer;">
+<a [routerLink]="[index]" routerLinkActive="app-active" class="metallic list-group-item clearfix" style="cursor: pointer;">
   <div class="pull-left">
     <h4 class="list-group-item-heading">{{ recipe.name }}</h4>
     <p class="list-group-item-text">{{ recipe.description }}</p>
