+++ src/app/recipes/recipe-list/recipe-list.component.css
@@ -1,5 +1,5 @@
 hr {
-  border-top: 1px solid gold;
+  border-top: 1px solid #e5bdf6;
 }
 
 .centered {
@@ -9,8 +9,11 @@ hr {
 }
 
 .centered>button {
-  background-color: gold;
-  border: 1px solid gold;
+  background-color: #e5bdf6;
   color: black;
   width: 75%;
+}
+
+.root {
+  max-height: 75vh;
 }
