+++ src/app/recipes/recipe-list/recipe-list.component.html
@@ -1,11 +1,11 @@
 
 <div class="row">
   <div class="col-xs-12 centered">
-    <button (click)="onNewRecipe()" class="btn btn-success">New Recipe</button>
+    <button (click)="onNewRecipe()" class="metallic btn">New Recipe</button>
   </div>
 </div>
 <hr>
-<div class="row">
+<div class="row root scrollable">
   <div class="col-xs-12" *ngFor="let recipe of recipes; let id = index">
     <app-recipe-item [recipe]="recipe" [index]="id"></app-recipe-item>
   </div>
