+++ src/app/recipes/recipe.model.ts
@@ -4,11 +4,13 @@ export class Recipe {
   public description: string;
   public imagePath: string;
   public ingredients: Ingredient[];
+  public instructions: string[];
 
-  constructor(name:string, desc: string, imagePath: string, ingredients: Ingredient[]) {
+  constructor(name:string, desc: string, imagePath: string, ingredients: Ingredient[] = [], instructions: string[] = []) {
     this.name = name;
     this.description = desc;
     this.imagePath = imagePath;
     this.ingredients = ingredients;
+    this.instructions = instructions;
   }
 }
