+++ src/app/recipes/recipes.component.ts
@@ -1,4 +1,3 @@
-import { RecipeService } from './../shared/services/recipe.service';
 import { Component, OnInit } from '@angular/core';
 
 @Component({
