+++ src/app/shopping/shopping-list-edit/shopping-list-edit.component.css
@@ -1,0 +1,11 @@
+hr {
+  border-top: 1px solid #e5bdf6;
+}
+
+label {
+  color: white;
+}
+
+.main {
+  width: 100% !important;
+}
