+++ src/app/shopping/shopping-list.component.css
@@ -1,0 +1,20 @@
+hr {
+  border-top: 1px solid #e5bdf6;
+}
+
+a {
+  width: 100%;
+}
+a:hover {
+  margin-left: 10px;
+}
+
+.black {
+  background-color: black;
+}
+
+.list-group {
+  overflow-y: auto;
+  overflow-x: hidden;
+  max-height: 65vh;
+}
