+++ src/app/shopping/shopping-list.component.html
@@ -2,8 +2,8 @@
   <div class="col-xs-10">
     <app-shopping-list-edit (onNewIngredient)="onNewIngredient($event)"></app-shopping-list-edit>
     <hr>
-    <ul class="list-group">
-      <a class="list-group-item"
+    <ul class="list-group scrollable">
+      <a class="list-group-item metallic"
          style="cursor: pointer"
          *ngFor="let ingredient of ingredients; let i = index"
          (click)="onEditItem(i)"
