+++ src/styles.css
@@ -1,9 +1,50 @@
 /* You can add global styles to this file, and also import other style files */
 .backdrop {
   /* background: linear-gradient(140deg, blue, green); */
-  background: black;
+  background-color: #000000;
+  background-image: linear-gradient(315deg, #000000 0%, #414141 74%);
   position: absolute;
   width: 100%;
   height: 100%;
   z-index: -5;
+}
+
+.metallic {
+  background-color: #d8dede;
+  background-image: linear-gradient(315deg, #d8dede 0%, #e5bdf6 74%);
+}
+
+.metallic-reverse {
+  background-color: #d8dede;
+  background-image: linear-gradient(315deg, #e5bdf6 74%, #d8dede);
+}
+
+.centered {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+}
+
+.success {
+  background-color: #80ff72;
+  background-image: linear-gradient(180deg, #80ff72 74%, #7ee8fa 0%);
+}
+
+.functional {
+  background-color: #009ffd;
+  background-image: linear-gradient(315deg, #009ffd 74%, #2a2a72 0%);
+  color: white;
+}
+
+.danger {
+  background-color: #f814149c;
+  background-image: linear-gradient(315deg, #d98324 74%, #a40606 0%);
+  color: white;
+}
+
+.scrollable {
+  overflow-x: hidden;
+  overflow-y: auto;
+  scrollbar-width: thin;
+  scrollbar-color: #e5bdf6;
 }
